:recycle: Heavy refactor of jwt/email routes/services
